add s:remove_from_buffer_list()
This functions checks if a given key actually exists in our
s:active_buffers dictionary, because there might be buffers that weren't
added to it in the first place.

At the moment this is the case for Vim's help pages and buffers that
aren't associated with any filename.

Thanks for reporting, Jeremy Mack.

Closes #1.