disable CVS detection by default
If you're working in a non-CVS repo and there are no changes, the repo
detection tries all supported repos in a certain order with CVS being
the last. If $CVSROOT is set that can take quite some time although
we're not even working in a CVS repo, so it's disabled by default now.

A new option for enabling CVS was added:

let g:signify_enable_cvs = 1