Fixed wrong jump if no count is specified
If jump next hunk with no count, it move last hunk of below hunks.

v:count step is 0, 2, 3, 4..., so if no count is specified, next hunk is -1(this is last hunk of below hunks).
Because of above reason, I change to use v:count1 instead of v:count.
v:count1 step is 1, 2, 3, 4... .