Make jumping behaviour more pragmatic
This patch makes jumps go the the first/last hunk when the given count
was bigger than the available hunks.

E.g. when there are only 2 hunks and you use 4]c, the cursor jumps to
the last hunk nevertheless. Before this patch the cursor wouldn't move
at all (which mimics the standard behaviour of ]c/[c).

Closes #41.