Dynamically re-apply diff options on every diff
Prior to this patch, it was not possible to dynamically apply arguments
to diff commands.  The g:signify_diffoptions variable was only read when
the plugin was loaded.

This patch changes the behavior so that before each diff it checks
g:signify_diffoptions to see what arguments should be passed to the diff
tool.

The motivating use case behind this change is to be able to diff against
a different branch in Git via a key mapping. However, This change will
also allow any other arbitrary diff options to be dynamically updated,
such as whitespace, etc.

Example: Show diff against master instead of against HEAD in Git
  nnoremap \u :let g:signify_diffoptions = { 'git': 'master'}<CR>:SignifyToggle<CR>