Revert "Escape 2>&1 in the p4 detection"
This reverts commit 5cdf72acd17d3559e4b5a207dea2993ddd77f50d.

`shellescape()`, which is what `sy#util#escape()` is a front-end to, is
only supposed to escape a single element of an argv list. In this case,
it's being used to escape multiple -- `p4 info 2>&1 >`. None of that
should be escaped anyway, because they need to be interpreted by the
shell.