Remove g:signify_vcs_list
The behaviour of how external VCS tools get registered changed a few times over
the lifetime of this plugin.

Back when Vim didn't have support for asynchronous execution, g:signify_vcs_list
was the most important option by far. If it wasn't set to only the VCS tools in
use, Sy would shell out for each tool it supported and Vim would block in the
meantime. Not cool.

Nowadays all the probing happens only once, in the background, and parallel. The
user won't notice any blocking.