Remove partial-change logic.
Previously signify handled a number of different cases for the
relationship between the number of additions and deletions. However this
wasn't really meaningful as it assumed that all of the added lines came
first, then the changed lines followed by the deletions. This
assumption has no relationship to the actual changes so it is better not
to have it.

Example diff with more additions:

    - apple
    + banana
    + pear

Previous behaviour:

    ! banana
    + pear

New behaviour:

    ! banana
    ! pear

Example diff with more subtractions:

    - apple
    - banana
    + pear

Previous behaviour: (uses changedelete which defaults to delete)

    * pear

New behaviour:

    ! pear