Override g:signify_sign_change_delete if it is too wide
Since the default config is a combination of g:signify_sign_change and
g:signify_sign_delete_first_line, it may be too wide for the sign
columns (either due to 'ambiwidth' being set to "double" or one of the
signs being customized to two characters).

Emit a diagnostic for why this is being overridden so there's at least a
hint of what's going on.

Closes #366